works on the jackal
