--- nav_goal_bridge/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/nav_goal_bridge
-[install]
-install_scripts=$base/lib/nav_goal_bridge
