--- nav_goal_bridge/setup.py
@@ -1,27 +1,0 @@
-from setuptools import setup
-
-package_name = 'nav_goal_bridge'
-
-setup(
-    name=package_name,
-    version='0.0.0',
-    packages=[package_name],
-    data_files=[
-        ('share/ament_index/resource_index/packages',
-            ['resource/' + package_name]),
-        ('share/' + package_name, ['package.xml']),
-    ],
-    install_requires=['setuptools'],
-    zip_safe=True,
-    maintainer='olagh',
-    maintainer_email='olagh@todo.todo',
-    description='TODO: Package description',
-    license='TODO: License declaration',
-    tests_require=['pytest'],
-    entry_points={
-        'console_scripts': [
-            'nav_goal_bridge_node = nav_goal_bridge.nav2_action_zmq:main'
-            'detect_eating_node = nav_goal_bridge.detect_eating:main'
-        ],
-    },
-)
