+++ shr_actions_cpp/src/find_person_action.cpp
@@ -9,8 +9,8 @@
 #include "shr_msgs/action/find_person_request.hpp"
 #include "shr_msgs/action/rotate_request.hpp"
 #include "shr_msgs/action/recognize_request.hpp"
-#include "nav2_msgs/action/navigate_to_pose.hpp"
-
+//#include "nav2_msgs/action/navigate_to_pose.hpp"
+#include "shr_msgs/action/navigate_to_pose.hpp"
 #include "shr_utils/utils.hpp"
 
 
@@ -38,7 +38,7 @@ namespace find_person_request {
                     std::bind(&FindPersonRequestActionServer::handle_cancel, this, _1),
                     std::bind(&FindPersonRequestActionServer::handle_accepted, this, _1));
 
-            navigation_action_client_ = rclcpp_action::create_client<nav2_msgs::action::NavigateToPose>(
+            navigation_action_client_ = rclcpp_action::create_client<shr_msgs::action::NavigateToPose>(
                     this, "navigate_to_pose");
             rotate_client_ = rclcpp_action::create_client<shr_msgs::action::RotateRequest>(
                     this, "rotate");
@@ -59,7 +59,7 @@ namespace find_person_request {
 
     private:
         rclcpp_action::Server<FindPersonRequest>::SharedPtr action_server_;
-        rclcpp_action::Client<nav2_msgs::action::NavigateToPose>::SharedPtr navigation_action_client_;
+        rclcpp_action::Client<shr_msgs::action::NavigateToPose>::SharedPtr navigation_action_client_;
         rclcpp_action::Client<shr_msgs::action::RotateRequest>::SharedPtr rotate_client_;
         rclcpp_action::Client<shr_msgs::action::RecognizeRequest>::SharedPtr recognize_face_client_;
