+++ shr_actions_py/package.xml
@@ -20,6 +20,7 @@
   <depend>tf2_geometry_msgs</depend>
   <depend>tf2_ros</depend>
   <depend>vlc</depend>
+  <depend>shr_msgs</depend>
 
   <exec_depend>rosidl_default_runtime</exec_depend>
