+++ shr_plan/launch/real_robot.launch.py
@@ -6,6 +6,7 @@ from launch.substitutions import LaunchConfiguration, PathJoinSubstitution
 from launch_ros.actions import Node
 from launch.conditions import IfCondition
 
+
 def generate_launch_description():
     smartthings_node = Node(
         package='smartthings_ros',
@@ -13,15 +14,27 @@ def generate_launch_description():
         output='screen'
     )
 
-    # nav_bridge = Node(
-    #     package='nav_goal_bridge',
-    #     executable='nav_goal_bridge_node',
-    #     output='screen'
-    # )
+    detect_eating_node = Node(
+        package='shr_world_state',
+        executable='detect_eating_real_node',
+        output='screen'
+    )
 
+    nav_bridge_cmd = Node(
+        package='shr_actions_py',
+        executable='nav2_zmq_action',
+        name='nav2_zmq_action',
+        output='screen')
+
+# TO DO ADD THE CAMERA TOPIC AS AN ARGUEMNT
     yolo_cmd = IncludeLaunchDescription(
         PythonLaunchDescriptionSource(PathJoinSubstitution([
             get_package_share_directory('yolostate'), 'launch', 'detecthuman.launch.py']))
+    )
+
+    yolo_house_cmd = IncludeLaunchDescription(
+        PythonLaunchDescriptionSource(PathJoinSubstitution([
+            get_package_share_directory('yolostate'), 'launch', 'detecthumaninhouserealcam.launch.py']))
     )
 
     tf_broadcast = Node(
@@ -31,9 +44,11 @@ def generate_launch_description():
     )
 
     ld = LaunchDescription()
+    ld.add_action(nav_bridge_cmd)
     ld.add_action(tf_broadcast)
     ld.add_action(smartthings_node)
     ld.add_action(yolo_cmd)
-    # ld.add_action(nav_bridge)
+    ld.add_action(yolo_house_cmd)
+    ld.add_action(detect_eating_node)
 
     return ld
