+++ shr_plan/src/actions/move_action_node.cpp
@@ -20,7 +20,8 @@
 
 #include "geometry_msgs/msg/pose.hpp"
 #include "geometry_msgs/msg/pose_with_covariance_stamped.hpp"
-#include "nav2_msgs/action/navigate_to_pose.hpp"
+//#include "nav2_msgs/action/navigate_to_pose.hpp"
+#include "shr_msgs/action/navigate_to_pose.hpp"
 #include "plansys2_executor/ActionExecutorClient.hpp"
 #include "rclcpp/rclcpp.hpp"
 #include "rclcpp_action/rclcpp_action.hpp"
@@ -31,106 +32,105 @@
 #include <shr_plan_parameters.hpp>
 
 
-
 namespace move_action {
-  using namespace std::chrono_literals;
+    using namespace std::chrono_literals;
 
-  class MoveAction : public plansys2::ActionExecutorClient {
-  public:
-    MoveAction(const std::string& action)
-        : plansys2::ActionExecutorClient(action, 500ms) {
+    class MoveAction : public plansys2::ActionExecutorClient {
+    public:
+        MoveAction(const std::string &action)
+                : plansys2::ActionExecutorClient(action, 500ms) {
 
-      set_parameter(rclcpp::Parameter("action_name", action));
-      tf_buffer_ =
-          std::make_unique<tf2_ros::Buffer>(this->get_clock());
-      tf_listener_ =
-          std::make_shared<tf2_ros::TransformListener>(*tf_buffer_);
+            set_parameter(rclcpp::Parameter("action_name", action));
+            tf_buffer_ =
+                    std::make_unique<tf2_ros::Buffer>(this->get_clock());
+            tf_listener_ =
+                    std::make_shared<tf2_ros::TransformListener>(*tf_buffer_);
 
-    }
+        }
 
-    void current_pos_callback(const geometry_msgs::msg::PoseWithCovarianceStamped::SharedPtr msg) {
-      current_pos_ = msg->pose.pose;
-    }
+        void current_pos_callback(const geometry_msgs::msg::PoseWithCovarianceStamped::SharedPtr msg) {
+            current_pos_ = msg->pose.pose;
+        }
 
-    rclcpp_lifecycle::node_interfaces::LifecycleNodeInterface::CallbackReturn
-    on_activate(const rclcpp_lifecycle::State &previous_state) {
-      send_feedback(0.0, "Move starting");
+        rclcpp_lifecycle::node_interfaces::LifecycleNodeInterface::CallbackReturn
+        on_activate(const rclcpp_lifecycle::State &previous_state) {
+            send_feedback(0.0, "Move starting");
 
-      navigation_action_client_ =
-          rclcpp_action::create_client<nav2_msgs::action::NavigateToPose>(
-              this,
-              "navigate_to_pose");
+            navigation_action_client_ =
+                    rclcpp_action::create_client<shr_msgs::action::NavigateToPose>(
+                            this,
+                            "navigate_to_pose");
 
-      bool is_action_server_ready = false;
-      do {
-        RCLCPP_INFO(get_logger(), "Waiting for /navigate_to_pose action server...");
+            bool is_action_server_ready = false;
+            do {
+                RCLCPP_INFO(get_logger(), "Waiting for /navigate_to_pose action server...");
 
-        is_action_server_ready =
-            navigation_action_client_->wait_for_action_server(std::chrono::seconds(5));
-      } while (!is_action_server_ready);
+                is_action_server_ready =
+                        navigation_action_client_->wait_for_action_server(std::chrono::seconds(5));
+            } while (!is_action_server_ready);
 
-      RCLCPP_INFO(get_logger(), "Navigation action server ready");
+            RCLCPP_INFO(get_logger(), "Navigation action server ready");
 
-      auto wp_to_navigate = get_arguments()[1];  // The goal is in the 3rd argument of the action
-      RCLCPP_INFO(get_logger(), "Start navigation to [%s]", wp_to_navigate.c_str());
+            auto wp_to_navigate = get_arguments()[1];  // The goal is in the 3rd argument of the action
+            RCLCPP_INFO(get_logger(), "Start navigation to [%s]", wp_to_navigate.c_str());
 
-      auto point = shr_utils::get_tf_as_point(*tf_buffer_, "map", wp_to_navigate);
-      dist_to_move = getDistance(point, current_pos_);
-      auto feedback_callback = [this](
-          shr_utils::NavigationGoalHandle::SharedPtr,
-          shr_utils::NavigationFeedback feedback) {
-        send_feedback(std::min(1.0, std::max(0.0, 1.0 - (feedback->distance_remaining / dist_to_move))),
-                      "Move running");
-      };
-      auto result_callback = [this](auto r) {
-      
-        finish(true, 1.0, "Move completed");
-      };
+            auto point = shr_utils::get_tf_as_point(*tf_buffer_, "map", wp_to_navigate);
+            dist_to_move = getDistance(point, current_pos_);
+            auto feedback_callback = [this](
+                    shr_utils::NavigationGoalHandle::SharedPtr,
+                    shr_utils::NavigationFeedback feedback) {
+                send_feedback(std::min(1.0, std::max(0.0, 1.0 - (feedback->distance_remaining / dist_to_move))),
+                              "Move running");
+            };
+            auto result_callback = [this](auto r) {
 
-      shr_utils::send_nav_request(*tf_buffer_, wp_to_navigate, now(),
-                                  navigation_action_client_, std::nullopt, feedback_callback, result_callback);
+                finish(true, 1.0, "Move completed");
+            };
 
-      return ActionExecutorClient::on_activate(previous_state);
-    }
+            shr_utils::send_nav_request(*tf_buffer_, wp_to_navigate, now(),
+                                        navigation_action_client_, std::nullopt, feedback_callback, result_callback);
 
-  private:
-    double getDistance(const geometry_msgs::msg::Pose &pos1, const geometry_msgs::msg::Pose &pos2) {
-      return sqrt(
-          (pos1.position.x - pos2.position.x) * (pos1.position.x - pos2.position.x) +
-          (pos1.position.y - pos2.position.y) * (pos1.position.y - pos2.position.y));
-    }
+            return ActionExecutorClient::on_activate(previous_state);
+        }
 
-    void do_work() {
-    }
+    private:
+        double getDistance(const geometry_msgs::msg::Pose &pos1, const geometry_msgs::msg::Pose &pos2) {
+            return sqrt(
+                    (pos1.position.x - pos2.position.x) * (pos1.position.x - pos2.position.x) +
+                    (pos1.position.y - pos2.position.y) * (pos1.position.y - pos2.position.y));
+        }
 
-    rclcpp_action::Client<nav2_msgs::action::NavigateToPose>::SharedPtr navigation_action_client_;
-    std::shared_future<shr_utils::NavigationGoalHandle::SharedPtr> future_navigation_goal_handle_;
-    shr_utils::NavigationGoalHandle::SharedPtr navigation_goal_handle_;
+        void do_work() {
+        }
 
-    std::shared_ptr<tf2_ros::TransformListener> tf_listener_{nullptr};
-    std::unique_ptr<tf2_ros::Buffer> tf_buffer_;
+        rclcpp_action::Client<shr_msgs::action::NavigateToPose>::SharedPtr navigation_action_client_;
+        std::shared_future<shr_utils::NavigationGoalHandle::SharedPtr> future_navigation_goal_handle_;
+        shr_utils::NavigationGoalHandle::SharedPtr navigation_goal_handle_;
 
-    rclcpp::Subscription<geometry_msgs::msg::PoseWithCovarianceStamped>::SharedPtr pos_sub_;
-    geometry_msgs::msg::Pose current_pos_;
-    nav2_msgs::action::NavigateToPose::Goal navigation_goal_;
+        std::shared_ptr<tf2_ros::TransformListener> tf_listener_{nullptr};
+        std::unique_ptr<tf2_ros::Buffer> tf_buffer_;
 
-    double dist_to_move;
-  };
+        rclcpp::Subscription<geometry_msgs::msg::PoseWithCovarianceStamped>::SharedPtr pos_sub_;
+        geometry_msgs::msg::Pose current_pos_;
+        shr_msgs::action::NavigateToPose::Goal navigation_goal_;
+
+        double dist_to_move;
+    };
 }
 
 int main(int argc, char **argv) {
-  rclcpp::init(argc, argv);
+    rclcpp::init(argc, argv);
 
-  auto parameter_node = std::make_shared<rclcpp::Node>("move_parameter_node");
-  auto param_listener = std::make_shared<shr_plan_parameters::ParamListener>(parameter_node);
-  auto params = param_listener->get_params();
+    auto parameter_node = std::make_shared<rclcpp::Node>("move_parameter_node");
+    auto param_listener = std::make_shared<shr_plan_parameters::ParamListener>(parameter_node);
+    auto params = param_listener->get_params();
 
-  auto node = std::make_shared<move_action::MoveAction>(params.move_robot_actions.actions[0]);
-  node->trigger_transition(lifecycle_msgs::msg::Transition::TRANSITION_CONFIGURE);
+    auto node = std::make_shared<move_action::MoveAction>(params.move_robot_actions.actions[0]);
+    node->trigger_transition(lifecycle_msgs::msg::Transition::TRANSITION_CONFIGURE);
 
-  rclcpp::spin(node->get_node_base_interface());
+    rclcpp::spin(node->get_node_base_interface());
 
-  rclcpp::shutdown();
+    rclcpp::shutdown();
 
-  return 0;
+    return 0;
 }
