+++ shr_plan/src/shr_plan_parameters.yaml
@@ -39,48 +39,48 @@ shr_plan_parameters:
     default_value: [ "bedroom", "kitchen", "couch", "door" ],
     description: "list of tf frame locations that the robot can visit",
   }
-#  tf_values:
-#    bedroom: {
-#      type: double_array,
-#      default_value: [ 11, 3.66, 0, 0, 0, 0, 1 ],
-#      description: "TF for kitchen",
-#    }
-#    kitchen: {
-#        type: double_array,
-#        default_value: [5.3, 2.5, 0, 0, 0, 0, 1 ],
-#        description: "TF for kitchen",
-#      }
-#    couch: {
-#      type: double_array,
-#      default_value: [ 1.8, -9, 0, 0, 0, 0, 1 ],
-#      description: "TF for couch",
-#      }
-#    door: {
-#      type: double_array,
-#      default_value: [ 2.75, 0.58, -7.76, 0, 0, 0, 1 ],
-#      description: "TF for couch",
-#      }
   tf_values:
     bedroom: {
       type: double_array,
-      default_value: [ 3.05, -1.4, -2.35, 0, 0, 0, 1 ],
+      default_value: [ 11, 3.66, 0, 0, 0, 0, 1 ],
       description: "TF for kitchen",
     }
     kitchen: {
         type: double_array,
-        default_value: [-2.38, -1.4, -0.92, 0, 0, 0, 1 ],
+        default_value: [5.3, 2.5, 0, 0, 0, 0, 1 ],
         description: "TF for kitchen",
       }
     couch: {
       type: double_array,
-      default_value: [-2.5, -1.4, -0.19, 0, 0, 0, 1 ],
+      default_value: [ 1.8, -9, 0, 0, 0, 0, 1 ],
       description: "TF for couch",
       }
     door: {
       type: double_array,
-      default_value: [ -1.28, -1.4, -3.25, 0, 0, 0, 1 ],
+      default_value: [ 2.75, 0.58, -7.76, 0, 0, 0, 1 ],
       description: "TF for couch",
       }
+#  tf_values:
+#    bedroom: {
+#      type: double_array,
+#      default_value: [ 3.05, -1.4, -2.35, 0, 0, 0, 1 ],
+#      description: "TF for kitchen",
+#    }
+#    kitchen: {
+#        type: double_array,
+#        default_value: [-2.38, -1.4, -0.92, 0, 0, 0, 1 ],
+#        description: "TF for kitchen",
+#      }
+#    couch: {
+#      type: double_array,
+#      default_value: [-2.5, -1.4, -0.19, 0, 0, 0, 1 ],
+#      description: "TF for couch",
+#      }
+#    door: {
+#      type: double_array,
+#      default_value: [ -1.28, -1.4, -3.25, 0, 0, 0, 1 ],
+#      description: "TF for couch",
+#      }
 
   patient_name: {
     type: string,
