+++ shr_world_state/setup.py
@@ -27,6 +27,8 @@ setup(
     license='TODO: License declaration',
     tests_require=['pytest'],
     entry_points={
-        'console_scripts': ['world_state_node = shr_world_state.world_state_node:main'],
+        'console_scripts': ['world_state_node = shr_world_state.world_state_node:main',
+                            'detect_eating_sim_node = shr_world_state.detect_eating_sim_node:main',
+                            'detect_eating_real_node = shr_world_state.detect_eating_real_node:main'],
     },
 )
