+++ yolostate/yolostate/detecthuman.py
@@ -22,10 +22,11 @@ ros2 run yolostate detecthuman --ros-args -p camera:=/smart_home/camera/color/im
 
 class DetectHuman(Node):
     def __init__(self):
-        super().__init__('detect_human')
+        super().__init__('detecthuman')
  
         self.declare_parameter('view_camera', True)
-        self.declare_parameter('camera', '/smart_home/camera/color/image_raw')
+        self.declare_parameter('camera', '/camera/color/image_raw')
+        # self.declare_parameter('camera', '/smart_home/camera/color/image_raw')
         self.declare_parameter('pub_human', '/detecthuman')
 
         param_camera_topic = self.get_parameter('camera').value
